requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
Pillow==9.3.0
requests==2.28.1
tk==0.1.0
urllib3==1.26.12
yarl==1.8.1
paho-mqtt==1.6.1
